Issue #4538: Modified CommentsIndentationCheckTest.java and moved its input files to the commentsindentation subdirectory
