+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/indentation/commentsindentation/InputCommentsIndentationNoNpe.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.puppycrawl.tools.checkstyle.checks.indentation;
+package com.puppycrawl.tools.checkstyle.checks.indentation.commentsindentation;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
